add 'ip rule' to busybox-full config.
This change adds the ability to mess with routing tables in the kernel.
This is useful for default routing all traffic into an openvpn tunnel without
mucking with the main routing table.  No static host route for the server,
and no changing the default gateway.  Just add a new table with a higher
prio that throws the server ip out and defaults into the vpn tunnel, like so:

$ export VPN_TBL=10
$ ip route add throw ${SERVER_IP} table ${VPN_TBL}
$ ip route add default via ${VPN_GW} dev tap0 table ${VPN_TBL}
$ ip rule add from all lookup ${VPN_TBL} prio 1000

And when the vpn is shut down:

$ ip route flush table ${VPN_TBL}

Neat, huh? ;-)

Signed-off-by: Jason <github@lakedaemon.net>